modularizing continued..
